Sweap unused under carpets (#6)
* WIP

* rm unused

* fix import

* fix conflict
